pyproject.toml
[project]
name = "trainscanner2"
version = "2.0"
description = ""
authors = [
    {name = "vitroid",email = "vitroid@gmail.com"}
]
readme = "README.md"
requires-python = ">=3.11,<3.14"

[tool.poetry.dependencies]
tiffeditor =">=0.4.0,<0.5.0"
pyperbox =">=0.1.1,<0.2.0"
pykalman =">=0.10.2,<0.11.0"
japanize-matplotlib =">=1.1.3,<2.0.0"
trainscanner = {path = "../TrainScanner"}
tiledimage = {path = "../TiledImage"}



[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
